projectcode.py
This portion of my code simply gets "permission" from football-data.org to access the trivia questions and does so using the API key I was able to get from the website itself. It also asks the user a couple of questions that were pulled from the websites question bank and gives feedback on the users answer (either right or wrong). Lastly, the users progress is saved as points and tallied up on the leaderboard.