Update projectcode.py
This code basically does the same thing as the previous commit, but I altered the code so that it would ask the user questions with the options as well as well as include a broader range of questions. This makes the program more straightforward to both the user and the programmer. And just like the previous program, the users correct responses as being saved and at the end of the trivia, the users final score is being displayed